requirements.txt
# Core Django
Django==5.2.6
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2

# Database
dj-database-url==3.0.1
psycopg2-binary==2.9.9
mssql-django==1.5
pyodbc==5.2.0

# PDF Generation
xhtml2pdf==0.2.17
reportlab==4.4.5
html5lib==1.1
pypdf==6.1.1
PyYAML==6.0.3
arabic-reshaper==3.0.0
python-bidi==0.6.6
pillow==10.4.0

# Excel
openpyxl==3.1.5

# Colombian Holidays
festivos-colombia==0.1.8

# Web Server
gunicorn==21.2.0
whitenoise==6.11.0

# Utils
requests==2.32.5
python-dateutil==2.9.0.post0
pytz==2025.2
